Provide scripts/import_event for reimporting a single org
Run as

    bundle exec rails runner scripts/import_event -- lrug

This will look through data/organisations.yml to find the org with the supplied
slug and run the reimport for that org. This is otherwise a copy and paste of
the organisation creation part of db/seeds.rb.

I found this useful while I was prepping the LRUG data, so I hope it's useful
for others.