Add involvements.yml for LRUG meetings
Open question here is that this is the _current_ list of organisers, and 2 of
them only started helping out from 2021 onwards. We've got a list of organisers
who were involved in our meetings before 2021, but no longer. We could add them
as a new "type" of organiser, but it'd link them to the whole series. I think
this is likely due to all the meetings for a single meetup being listed as a
single event so there's only space for one involvements.yml. The flip side of
course being that creating a folder + event.yml, involvements.yml, videos.yml
for each meeting is probably also onerous.

Perhaps we just need to allow for some date ranges in the involvements.yml and
to enrich the backfill script.